web/requirements.txt
Flask==0.11.1
python-chess==0.15.3
pystockfish==1.0.1